rename preferences saving to api
